requirements.txt
streamlit
Pillow
cryptography